+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
+- Fix MapView::getBoundingBox(): add map-center as offset
 - Update java-gtk to 0.6.0 (or 0.6.1)
 - Update MapsForge to 0.24.1
 - Update release plugin to 1.18.18